+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- Filter `cdylib` libraries by the `TargetKind` and not by the file extension.
 
 ## [0.5.0] - 2018-05-29
 ### Added